Class refactoring.
